Added basic support for RAM/CPU/FSSize data
Currenty the app will display the current CPU load, RAM in use, and the percentage used of each mounted drive.

Also, if you execute the script as `node .\Apollo.js pollRate` where pollRate is the rate in ms you want it to update, it will update at that interval. Otherwise, it updates every 5 seconds.